+++ inc/user.class.php
@@ -311,46 +311,29 @@ class User extends CommonDBTM {
 
 
    function cleanDBonPurge() {
+
       global $DB;
 
-      $DB->delete(
-         'glpi_profiles_users', [
-            'users_id' => $this->fields['id']
-         ]
-      );
+      // ObjectLock does not extends CommonDBConnexity
+      $ol = new ObjectLock();
+      $ol->deleteByCriteria(['users_id' => $this->fields['id']]);
 
-      if ($this->fields['id'] > 0) { // Security
-         $DB->delete(
-            'glpi_displaypreferences', [
-               'users_id' => $this->fields['id']
-            ]
-         );
-
-         $DB->delete(
-            'glpi_savedsearches_users', [
-               'users_id' => $this->fields['id']
-            ]
-         );
-      }
-
-      // Delete own reminders
-      $DB->delete(
-         'glpi_reminders', [
-            'users_id' => $this->fields['id']
-         ]
-      );
+      // Reminder does not extends CommonDBConnexity
+      $r = new Reminder();
+      $r->deleteByCriteria(['users_id' => $this->fields['id']]);
 
       // Delete private bookmark
-      $DB->delete(
-         'glpi_savedsearches', [
-            'users_id'     => $this->fields['id'],
-            'is_private'   => 1
+      $ss = new SavedSearch();
+      $ss->deleteByCriteria(
+         [
+            'users_id'   => $this->fields['id'],
+            'is_private' => 1,
          ]
       );
 
       // Set no user to public bookmark
       $DB->update(
-         'glpi_savedsearches', [
+         SavedSearch::getTable(), [
             'users_id' => 0
          ], [
             'users_id' => $this->fields['id']
@@ -369,40 +352,29 @@ class User extends CommonDBTM {
          ]
       );
 
-      $gu = new Group_User();
-      $gu->cleanDBonItemDelete($this->getType(), $this->fields['id']);
-
-      $tu = new Ticket_User();
-      $tu->cleanDBonItemDelete($this->getType(), $this->fields['id']);
-
-      $pu = new Problem_User();
-      $pu->cleanDBonItemDelete($this->getType(), $this->fields['id']);
-
-      $cu = new Change_User();
-      $cu->cleanDBonItemDelete($this->getType(), $this->fields['id']);
-
-      $DB->delete(
-         'glpi_projecttaskteams', [
-            'items_id'  => $this->fields['id'],
-            'itemtype'  => __CLASS__
+      $this->deleteChildrenAndRelationsFromDb(
+         [
+            Certificate_Item::class,
+            Change_User::class,
+            Group_User::class,
+            KnowbaseItem_User::class,
+            Problem_User::class,
+            Profile_User::class,
+            ProjectTaskTeam::class,
+            ProjectTeam::class,
+            Reminder_User::class,
+            RSSFeed_User::class,
+            SavedSearch_User::class,
+            Ticket_User::class,
+            UserEmail::class,
          ]
       );
 
-      $DB->delete(
-         'glpi_projectteams', [
-            'items_id'  => $this->fields['id'],
-            'itemtype'  => __CLASS__
-         ]
-      );
-
-      $kiu = new KnowbaseItem_User();
-      $kiu->cleanDBonItemDelete($this->getType(), $this->fields['id']);
-
-      $ru = new Reminder_User();
-      $ru->cleanDBonItemDelete($this->getType(), $this->fields['id']);
-
-      $ue = new UserEmail();
-      $ue->deleteByCriteria(['users_id' => $this->fields['id']]);
+      if ($this->fields['id'] > 0) { // Security
+         // DisplayPreference does not extends CommonDBConnexity
+         $dp = new DisplayPreference();
+         $dp->deleteByCriteria(['users_id' => $this->fields['id']]);
+      }
 
       $this->dropPictureFiles($this->fields['picture']);